Add shutdownDelay (Due to Helidon bug)
There is a regression bug in Helidon that means it isn't performing a graceful shutdown.
Refer: https://github.com/helidon-io/helidon/issues/10156

A less than ideal workaround is to add in a shutdown delay. This will not be required once a fix is applied in Helidon, but we may want to keep this shutdownDelay support just in case.