Fix the shutdownDelay to occur after status = STOPPING  (#193)
* Default shutdownDelay to 200 millis (Due to Helidon bug)

There is a regression bug in Helidon that means it isn't performing a graceful shutdown.
Refer: https://github.com/helidon-io/helidon/issues/10156

The 200ms default shutdown delay is a workaround, will change this to default 0 when a Helidon fix is released.

* Version 1.5-RC2

* Fix the shutdownDelay to occur after status = STOPPING

In K8s we need the delay after the actual status has been set to STOPPING. This fixes that.